+++ lib/src/require.dart
@@ -7,9 +7,10 @@ class Require {
 
   /// Throws an exception with [message] if [check] is [false]
   /// `since 0.7.0`
-  static void require(bool check, String message) {
+  static bool require(bool check, String message) {
     if (check.isFalse) {
       throw Exception(message);
     }
+    return check;
   }
 }