+++ lib/pouch.dart
@@ -16,5 +16,6 @@ export 'src/objects.dart';
 export 'src/predicates.dart';
 export 'src/require.dart';
 export 'src/stream_extensions.dart';
+export 'src/string_extensions.dart';
 export 'src/strings.dart';
 export 'src/wait.dart';